pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scue"
version = "0.1.0"
description = "Automated lighting/laser/visual cue generation for live DJ sets"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=11.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
analysis = [
    "librosa>=0.10.0",
    "ruptures>=1.1.0",
]
usb = [
    "rbox>=0.1.7",          # exportLibrary.db reading (OneLibrary/DLP)
    "pyrekordbox>=0.4.4",   # ANLZ file parsing (beat grid, cues, phrases)
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["scue*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
